chore(registration-react): separate filter-pills as component
