+++ applications/registration-react/src/pages/dataset-registration-page/form-los/filter-pills-los/filter-pills-los.scss
@@ -1,0 +1,10 @@
+@import '../../../../../node_modules/designsystemet/fdk-designsystem-bootstrap4/scss/colors.scss';
+
+.fdk-filter-pill {
+  cursor: pointer;
+  &:after {
+    content: '\2715';
+    color: $color-white;
+    margin-left: 8px;
+  }
+}
