+++ applications/registration-react/src/pages/dataset-registration-page/form-los/form-los.component.jsx
@@ -5,6 +5,7 @@ import { Field } from 'redux-form';
 import localization from '../../../lib/localization';
 import Helptext from '../../../components/helptext/helptext.component';
 import { FieldTreeLos } from './field-tree-los/field-tree-los.component';
+import { FilterPillsLos } from './filter-pills-los/filter-pills-los.component';
 
 export const FormLOS = ({ losItems }) => {
   if (!losItems) {
@@ -14,6 +15,11 @@ export const FormLOS = ({ losItems }) => {
     <form>
       <div className="form-group">
         <Helptext title={localization.schema.los.helptext.title} />
+        <Field
+          name="theme"
+          component={FilterPillsLos}
+          losItems={losItems}
+        />
         <Field
           name="theme"
           component={FieldTreeLos}
